refactor: Separate major logical units out of main
